requirements.txt
smolagents[tools]
duckduckgo-search
python-dotenv
smolagents[telemetry]
opentelemetry-sdk
openinference-instrumentation-smolagents
opentelemetry-exporter-otlp